prelim changes with Kmoore
